优化功能

+++ composeApp/src/desktopMain/kotlin/sophon/desktop/core/Context.kt
@@ -68,43 +68,49 @@ object Context {
         }
     }
 
-    fun selectDevice(deviceName: String) {
+    fun selectDevice(deviceName: String, pushDex: Boolean = false) {
         scope.launch {
             if (deviceName.isBlank()) {
                 _stream.update { state -> state.copy(status = State.Status.Success("未找到连接设备，直接进入主页")) }
                 return@launch
             }
 
-            val dexInDevice =
-                "adb shell find $SERVER_DST_PATH".simpleShell().trimEnd().split("/").last()
+            if (pushDex) {
+                val dexInDevice =
+                    "adb shell find $SERVER_DST_PATH".simpleShell().trimEnd().split("/").last()
 
-            if (dexInDevice != SERVER_SRC_DEX_NAME) {
-                _stream.update { state -> state.copy(status = State.Status.Loading("监测组件版本不一致，正在重新推送")) }
-                //如果手机里dex文件和当前文件不一样，就重新推送并启动
-                "adb shell rm -rf $SERVER_DST_DIR".simpleShell()
-                "adb push $SERVER_SRC_PATH $SERVER_DST_PATH".simpleShell()
+                if (dexInDevice != SERVER_SRC_DEX_NAME) {
+                    _stream.update { state -> state.copy(status = State.Status.Loading("监测组件版本不一致，正在重新推送")) }
+                    //如果手机里dex文件和当前文件不一样，就重新推送并启动
+                    "adb shell rm -rf $SERVER_DST_DIR".simpleShell()
+                    "adb push $SERVER_SRC_PATH $SERVER_DST_PATH".simpleShell()
+                }
+
+                _stream.update { state -> state.copy(status = State.Status.Loading("正在启动监测组件")) }
+                val pidFile = File("/data/local/tmp/sophon/sophon_pid.txt")
+                if (pidFile.parentFile.exists().not()) {
+                    pidFile.parentFile.mkdir()
+                }
+
+                // 读取PID文件获取进程ID
+                val lastPid =
+                    runCatching { "adb shell cat $pidFile".simpleShell().trim() }.getOrElse { "" }
+
+                //停止当前server进程
+                "adb shell kill $lastPid".simpleShell()
+                println("停止当前server进程: $lastPid")
+
+                // 启动进程并保存PID
+                "adb shell \"export CLASSPATH=${SERVER_DST_PATH};nohup app_process $SERVER_DST_DIR $SERVER_MAIN_CLASS > /dev/null 2>&1 & echo \\$! > ${pidFile.absolutePath}\"".simpleShell()
+                delay(500)
+
+                // 读取PID文件获取进程ID
+                val pid =
+                    runCatching { "adb shell cat $pidFile".simpleShell().trim() }.getOrElse { "" }
+                println("启动的应用进程ID: $pid")
+
+                SocketClient.connect(deviceName)
             }
-
-            _stream.update { state -> state.copy(status = State.Status.Loading("正在启动监测组件")) }
-            val pidFile = File("/data/local/tmp/sophon/sophon_pid.txt")
-            if (pidFile.parentFile.exists().not()){
-                pidFile.parentFile.mkdir()
-            }
-
-            // 读取PID文件获取进程ID
-            val lastPid = runCatching { "adb shell cat $pidFile".simpleShell().trim() }.getOrElse { "" }
-
-            //停止当前server进程
-            "adb shell kill $lastPid".simpleShell()
-            println("停止当前server进程: $lastPid")
-
-            // 启动进程并保存PID
-            "adb shell \"export CLASSPATH=${SERVER_DST_PATH};nohup app_process $SERVER_DST_DIR $SERVER_MAIN_CLASS > /dev/null 2>&1 & echo \\$! > ${pidFile.absolutePath}\"".simpleShell()
-            delay(500)
-
-            // 读取PID文件获取进程ID
-            val pid = runCatching { "adb shell cat $pidFile".simpleShell().trim() }.getOrElse { "" }
-            println("启动的应用进程ID: $pid")
 
             _stream.update { state ->
                 state.copy(
@@ -112,7 +118,6 @@ object Context {
                     selectedDevice = deviceName
                 )
             }
-            SocketClient.connect(deviceName)
         }
     }
